# src/fire/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Firestore API overview",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080/api/v1"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "operationId": "listVersionsv1",
        "summary": "List API versions",
        "responses": {
          "200": {
            "description": "200 response",
            "content": {
              "application/json": {
                "examples": {
                  "foo": {
                    "value": {
                      "versions": [
                        {
                          "status": "CURRENT",
                          "updated": "2020-01-05T11:33:21Z",
                          "id": "v1.0",
                          "links": [
                            {
                              "href": "http://127.0.0.1:8080/api/v1/",
                              "rel": "self"
                            }
                          ]
                        },
                        {
                          "status": "EXPERIMENTAL",
                          "updated": "2020-01-05T11:33:21Z",
                          "id": "v2.0",
                          "links": [
                            {
                              "href": "http://127.0.0.1:8080/api/v2/",
                              "rel": "self"
                            }
                          ]
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/fire/": {
      "get": {
        "operationId": "getTopCollectionsv1",
        "summary": "Get top-level collections",
        "responses": {
          "200": {
            "description": "200 response",
            "content": {
              "application/json": {
                "examples": {
                  "foo": {
                    "value": []
                  }
                }
              }
            }
          }
        }
      }
    },
    "/fire/{coll}/": {
      "get": {
        "operationId": "getTopCollectionv1",
        "summary": "Get documents in top-level collection",
        "parameters": [
          {
            "name": "coll",
            "in": "path",
            "description": "Name of the top-level collection",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Get page result (default 1)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort field (default none)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fields",
            "in": "query",
            "description": "Get comma-separated list of fields (default all truncated)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "content": {
              "application/json": {
                "examples": {
                  "foo": {
                    "value": []
                  }
                }
              }
            }
          }
        }
      }
    },
    "/fire/{coll}/{ref}": {
      "get": {
        "operationId": "getDocumentv1",
        "summary": "Get document",
        "parameters": [
          {
            "name": "coll",
            "in": "path",
            "description": "Name of the top-level collection",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ref",
            "in": "path",
            "description": "ID of the document in the collection",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fields",
            "in": "query",
            "description": "Get comma-separated list of fields (default all)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "children",
            "in": "query",
            "description": "Get subcollections of document (default false)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "content": {
              "application/json": {
                "examples": {
                  "foo": {
                    "value": {}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/fire/{coll}/{ref}/{subcoll}/": {
      "get": {
        "operationId": "getSubCollectionv1",
        "summary": "Get documents in subcollection",
        "parameters": [
          {
            "name": "coll",
            "in": "path",
            "description": "Name of the top-level collection",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ref",
            "in": "path",
            "description": "ID of the document in the collection",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subcoll",
            "in": "path",
            "description": "Name of the subcollection",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Get page result (default 1)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort field (default none)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fields",
            "in": "query",
            "description": "Get comma-separated list of fields (default all truncated)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "children",
            "in": "query",
            "description": "Get subcollections of document (default false)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "content": {
              "application/json": {
                "examples": {
                  "foo": {
                    "value": []
                  }
                }
              }
            }
          }
        }
      }
    },
    "/fire/{coll}/{ref}/{subcoll}/{subref}": {
      "get": {
        "operationId": "getSubDocumentv1",
        "summary": "Get document in subcollection",
        "parameters": [
          {
            "name": "coll",
            "in": "path",
            "description": "Name of the top-level collection",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ref",
            "in": "path",
            "description": "ID of the document in the collection",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subcoll",
            "in": "path",
            "description": "Name of the subcollection",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subref",
            "in": "path",
            "description": "ID of the document in the subcollection",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fields",
            "in": "query",
            "description": "Get comma-separated list of fields (default all truncated)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "content": {
              "application/json": {
                "examples": {
                  "foo": {
                    "value": []
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}